+++ backend/app/db/models/expediente.py
@@ -1,6 +1,6 @@
 from typing import List
 from datetime import datetime
-from sqlalchemy import Integer, String, DateTime, UniqueConstraint
+from sqlalchemy import BigInteger, String, DateTime, UniqueConstraint
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from .base import Base
 from .expediente_documento import T_Expediente_Documento
@@ -9,7 +9,7 @@ class T_Expediente(Base):
     __tablename__ = "T_Expediente"
     __table_args__ = (UniqueConstraint("CT_Num_expediente", name="uq_expediente_num"),)
 
-    CN_Id_expediente: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
+    CN_Id_expediente: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
     CT_Num_expediente: Mapped[str] = mapped_column(String(60), nullable=False)
     CF_Fecha_creacion: Mapped[datetime] = mapped_column(DateTime, nullable=False, default=datetime.utcnow)
